requirements.txt
accelerate==1.8.1
annotated-types==0.7.0
anyio==4.9.0
av==14.4.0
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
diffusers==0.34.0
fastapi==0.115.14
filelock==3.18.0
fsspec==2025.5.1
h11==0.16.0
hf-xet==1.1.5
huggingface-hub==0.33.1
idna==3.10
importlib_metadata==8.7.0
Jinja2==3.1.6
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.5
numpy==2.3.1
opencv-python==4.11.0.86
packaging==25.0
pillow==11.2.1
psutil==7.0.0
pydantic==2.11.7
pydantic_core==2.33.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.4
safetensors==0.5.3
setuptools==80.9.0
sniffio==1.3.1
starlette==0.46.2
sympy==1.14.0
tokenizers==0.21.2
torch==2.7.1
torchvision==0.22.1
tqdm==4.67.1
transformers==4.53.0
typing-inspection==0.4.1
typing_extensions==4.14.0
urllib3==2.5.0
uvicorn==0.35.0
zipp==3.23.0
